RI-969354 Выделил модуль под uikit
