--- app/src/main/java/ru/rollinitiative/ui/element/base/UiKitAppearance.kt
@@ -1,6 +1,0 @@
-package ru.rollinitiative.ui.element.base
-
-/**
- * Маркерный интерфейс сущности, которая предоставляет значения атрибутов оформления компонента.
- */
-interface UiKitAppearance
