--- app/src/main/java/ru/rollinitiative/ui/element/base/UiKitSize.kt
@@ -1,6 +1,0 @@
-package ru.rollinitiative.ui.element.base
-
-/**
- * Маркерный интерфейс сущности, которая предоставляет размерности компонентов и числовые значения.
- */
-interface UiKitSize
