--- app/src/main/java/ru/rollinitiative/ui/theme/UiKitIcons.kt
@@ -1,3 +1,0 @@
-package ru.rollinitiative.ui.theme
-
-object UiKitIcons
