+++ core-ui/compose/uikit/build.gradle.kts
@@ -1,0 +1,8 @@
+plugins {
+    id("ru.rollinitiative.android.library.plugin")
+    id("ru.rollinitiative.compose.plugin")
+}
+
+android {
+    namespace = "ru.rollinitiative.core.ui.compose.uikit"
+}
