+++ core-ui/compose/uikit/src/main/kotlin/ru/rollinitiative/core/ui/compose/uikit/element/base/UiKitAppearance.kt
@@ -1,0 +1,6 @@
+package ru.rollinitiative.core.ui.compose.uikit.element.base
+
+/**
+ * Маркерный интерфейс сущности, которая предоставляет значения атрибутов оформления компонента.
+ */
+interface UiKitAppearance
