+++ core-ui/compose/uikit/src/main/kotlin/ru/rollinitiative/core/ui/compose/uikit/element/base/UiKitSize.kt
@@ -1,0 +1,6 @@
+package ru.rollinitiative.core.ui.compose.uikit.element.base
+
+/**
+ * Маркерный интерфейс сущности, которая предоставляет размерности компонентов и числовые значения.
+ */
+interface UiKitSize
