+++ core-ui/compose/uikit/src/main/kotlin/ru/rollinitiative/core/ui/compose/uikit/element/button/regular/preview/RegularButtonPreviewInfo.kt
@@ -1,0 +1,11 @@
+package ru.rollinitiative.core.ui.compose.uikit.element.button.regular.preview
+
+import ru.rollinitiative.core.ui.compose.uikit.element.button.regular.RegularButtonAppearance
+import ru.rollinitiative.core.ui.compose.uikit.element.button.regular.RegularButtonSize
+
+internal data class RegularButtonPreviewInfo(
+    val useDarkMode: Boolean,
+    val size: RegularButtonSize,
+    val isEnabled: Boolean = true,
+    val appearance: RegularButtonAppearance
+)
