+++ core-ui/compose/uikit/src/main/kotlin/ru/rollinitiative/core/ui/compose/uikit/element/button/regular/preview/RegularButtonPreviewParameterProvider.kt
@@ -1,0 +1,28 @@
+package ru.rollinitiative.core.ui.compose.uikit.element.button.regular.preview
+
+import androidx.compose.ui.tooling.preview.PreviewParameterProvider
+import ru.rollinitiative.core.ui.compose.uikit.element.button.regular.RegularButtonAppearances
+import ru.rollinitiative.core.ui.compose.uikit.element.button.regular.RegularButtonSizes
+
+internal class RegularButtonPreviewParameterProvider :
+    PreviewParameterProvider<RegularButtonPreviewInfo> {
+    override val values: Sequence<RegularButtonPreviewInfo>
+        get() {
+            return sequence {
+                val sizes = listOf(RegularButtonSizes.Medium, RegularButtonSizes.Small)
+                val appearances = listOf(RegularButtonAppearances.Primary)
+
+                for (size in sizes) {
+                    for (appearance in appearances) {
+                        yield(
+                            RegularButtonPreviewInfo(
+                                size = size,
+                                appearance = appearance,
+                                useDarkMode = true,
+                            )
+                        )
+                    }
+                }
+            }
+        }
+}
