+++ core-ui/compose/uikit/src/main/kotlin/ru/rollinitiative/core/ui/compose/uikit/theme/UiKitIcons.kt
@@ -1,0 +1,3 @@
+package ru.rollinitiative.core.ui.compose.uikit.theme
+
+object UiKitIcons
