+++ core-ui/compose/uikit/src/main/res/font/inter_regular.otf
